feat(dashboard): design updates to the install details
