+++ services/dashboard-ui/client/views/install/routes.tsx
@@ -17,6 +17,7 @@ import { ActionRunLayout } from './ActionRunLayout'
 import { ActionRunDetail } from './ActionRunDetail'
 import { ActionRunLogsPage } from './ActionRunLogs'
 import { SandboxRunDetail } from './SandboxRunDetail'
+import { CurrentInputs } from './CurrentInputs'
 import { ViewState } from './ViewState'
 import { WorkflowDetail } from './WorkflowDetail'
 import { RunnerJobDetail } from './RunnerJobDetail'
@@ -34,6 +35,7 @@ export const installRoutes: RouteObject[] = [
       { path: ':orgId/installs/:installId/roles', element: <Roles /> },
       { path: ':orgId/installs/:installId/policies', element: <Policies /> },
       { path: ':orgId/installs/:installId/runner', element: <Runner /> },
+      { path: ':orgId/installs/:installId/inputs', element: <CurrentInputs /> },
       { path: ':orgId/installs/:installId/state', element: <ViewState /> },
       {
         path: ':orgId/installs/:installId/runner/jobs/:jobId',
